+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Possibility to connect a non-redundant stream to a redundant one (one of the pair)
 - Milan GetStreamInfo extended information
 - Detection of devices not supporting _Acquire_ and/or _Lock_ commands
+- Display of the current dynamic mappings without having to edit them, in StreamPort descriptor
+- Button to clear all dynamic mappings in StreamPort descriptor
 
 ### Changed
 - Icon when an entity is acquired by Hive (changed color from orange to green)